+++ backend/http/router.go
@@ -11,6 +11,8 @@ import (
 
 	"github.com/gtsteffaniak/filebrowser/settings"
 	"github.com/gtsteffaniak/filebrowser/storage"
+
+	httpSwagger "github.com/swaggo/http-swagger" // http-swagger middleware
 )
 
 // Embed the files in the frontend/dist directory
@@ -72,45 +74,44 @@ func StartHttp(Service ImgService, storage *storage.Storage, cache FileCache) {
 	// API group routing
 	api := http.NewServeMux()
 
+	// User routes
 	api.HandleFunc("GET /users", withAdmin(usersGetHandler))
 	api.HandleFunc("POST /users", withAdmin(usersPostHandler))
 	api.HandleFunc("GET /users/{id}", withSelfOrAdmin(userGetHandler))
 	api.HandleFunc("PUT /users/{id}", withSelfOrAdmin(userPutHandler))
-	//api.HandleFunc("POST /users/{id}", withSelfOrAdmin(userPostHandler))
 	api.HandleFunc("DELETE /users/{id}", withSelfOrAdmin(userDeleteHandler))
 
-	// API routes
+	// Auth routes
 	api.HandleFunc("POST /login", loginHandler)
 	api.HandleFunc("GET /signup", signupHandler)
 	api.HandleFunc("POST /renew", withUser(renewHandler))
+
 	// Resources routes
 	api.HandleFunc("GET /resources/", withUser(resourceGetHandler))
 	api.HandleFunc("DELETE /resources/", withUser(resourceDeleteHandler))
 	api.HandleFunc("POST /resources/", withUser(resourcePostHandler))
 	api.HandleFunc("PUT /resources/", withUser(resourcePutHandler))
 	api.HandleFunc("PATCH /resource/", withUser(resourcePatchHandler))
-
-	// Additional API routes
 	api.HandleFunc("GET /usage/", withUser(diskUsage))
+	api.HandleFunc("GET /raw", withUser(rawHandler))
+	api.HandleFunc("GET /preview/{size}/{path...}", withUser(previewHandler))
 
+	// Share routes
 	api.HandleFunc("GET /shares", withPermShare(shareListHandler))
-
 	api.HandleFunc("GET /share/", withPermShare(shareGetsHandler))
 	api.HandleFunc("POST /share/", withPermShare(sharePostHandler))
 	api.HandleFunc("DELETE /share/", withPermShare(shareDeleteHandler))
 
-	api.HandleFunc("GET /settings", withAdmin(settingsGetHandler))
-	api.HandleFunc("PUT /settings", withAdmin(settingsPutHandler))
-
-	api.HandleFunc("GET /raw", withUser(rawHandler))
-
-	api.HandleFunc("GET /preview/{size}/{path...}", withUser(previewHandler))
-
-	api.HandleFunc("GET /search/", withUser(searchHandler))
-
+	// Public routes
 	api.HandleFunc("GET /public/publicUser", publicUserGetHandler)
 	api.HandleFunc("GET /public/dl", withHashFile(publicDlHandler))
 	api.HandleFunc("GET /public/share/", withHashFile(publicShareHandler))
+
+	// Settings routes
+	api.HandleFunc("GET /settings", withAdmin(settingsGetHandler))
+	api.HandleFunc("PUT /settings", withAdmin(settingsPutHandler))
+
+	api.HandleFunc("GET /search", withUser(searchHandler))
 
 	router.Handle("/api/", http.StripPrefix("/api", api))
 
@@ -120,6 +121,16 @@ func StartHttp(Service ImgService, storage *storage.Storage, cache FileCache) {
 
 	// health
 	router.HandleFunc("GET /health", healthHandler)
+
+	// Swagger
+	router.Handle("/swagger/",
+		httpSwagger.Handler(
+			httpSwagger.URL("/swagger/doc.json"), //The url pointing to API definition
+			httpSwagger.DeepLinking(true),
+			httpSwagger.DocExpansion("none"),
+			httpSwagger.DomID("swagger-ui"),
+		),
+	)
 
 	log.Printf("listing on port: %d", config.Server.Port)
 	err = http.ListenAndServe(fmt.Sprintf(":%d", config.Server.Port), muxWithMiddleware(router))