+++ CHANGELOG.md
@@ -7,14 +7,16 @@ All notable changes to this project will be documented in this file. For commit 
   **New Features**:
   - Create long-live api tokens to interact with API.
   - Helpful swagger page for API usage.
+  - Add file size to search preview.
 
   **Notes**:
-  - Replaced backend http framework with go standard library
+  - Replaced backend http framework with go standard library.
   - Right-click Context menu can target the item that was right-clicked.
   - adjusted settings menu for mobile, always shows all available cards rather than grayed out cards that need to be clicked.
 
   **Bugfixes**:
-  - Fixed ui bug with shares with password
+  - Fixed ui bug with shares with password.
+  - Fix empty directory load issue.
 
 ## v0.2.10